+++ .gitignore
@@ -1,3 +1,6 @@
+__version__.py
+
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
