+++ .vscode/settings.json
@@ -6,9 +6,13 @@
         "-p",
         "test_*.py"
     ],
-    "python.testing.pytestEnabled": false,
-    "python.testing.unittestEnabled": true,
+    "python.testing.pytestEnabled": true,
+    "python.testing.unittestEnabled": false,
     "python.linting.flake8Args": [
         "--config=pyproject.toml"
+    ],
+    "python-envs.pythonProjects": [],
+    "python.testing.pytestArgs": [
+        "tests"
     ]
 }
