+++ pyproject.toml
@@ -48,3 +48,4 @@ column_limit = 127
 [tool.setuptools_scm]
 version_scheme = "post-release"
 local_scheme = "node-and-date"
+write_to = "dicomfix/__version__.py"
